requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.36
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.3.0
python-multipart==0.0.9
jinja2==3.1.4
aiosqlite==0.20.0
python-dateutil==2.9.0
httpx==0.27.0
